Refactored
Created `Buffer class.h`, `Cursor class.h`, `Auxiliary functions.h` with declaration of the functional and .cpp files definition.

Initially was created a new project with new files in a way it is present now then it all was pasted to the repo, all previous files were deleted(except png. tests).

The logic has not been changed.
