+++ C Text Editor.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.34928.147
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "TestProj", "TestProj.vcxproj", "{3052527B-F682-465B-A93A-2259CE2B9193}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3052527B-F682-465B-A93A-2259CE2B9193}.Debug|x64.ActiveCfg = Debug|x64
+		{3052527B-F682-465B-A93A-2259CE2B9193}.Debug|x64.Build.0 = Debug|x64
+		{3052527B-F682-465B-A93A-2259CE2B9193}.Debug|x86.ActiveCfg = Debug|Win32
+		{3052527B-F682-465B-A93A-2259CE2B9193}.Debug|x86.Build.0 = Debug|Win32
+		{3052527B-F682-465B-A93A-2259CE2B9193}.Release|x64.ActiveCfg = Release|x64
+		{3052527B-F682-465B-A93A-2259CE2B9193}.Release|x64.Build.0 = Release|x64
+		{3052527B-F682-465B-A93A-2259CE2B9193}.Release|x86.ActiveCfg = Release|Win32
+		{3052527B-F682-465B-A93A-2259CE2B9193}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {91F1D293-23CA-4608-8BEA-A11FB4D7B1D3}
+	EndGlobalSection
+EndGlobal
