--- Source/Source.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34728.123
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Source", "Source.vcxproj", "{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}.Debug|x64.ActiveCfg = Debug|x64
-		{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}.Debug|x64.Build.0 = Debug|x64
-		{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}.Debug|x86.ActiveCfg = Debug|Win32
-		{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}.Debug|x86.Build.0 = Debug|Win32
-		{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}.Release|x64.ActiveCfg = Release|x64
-		{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}.Release|x64.Build.0 = Release|x64
-		{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}.Release|x86.ActiveCfg = Release|Win32
-		{3A72B2F8-DA95-4B32-B0AD-AF31C87523F0}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C50779A5-A76F-4EA9-9A50-F810048B2F20}
-	EndGlobalSection
-EndGlobal
