+++ TestProj.vcxproj
@@ -21,9 +21,10 @@
   <PropertyGroup Label="Globals">
     <VCProjectVersion>17.0</VCProjectVersion>
     <Keyword>Win32Proj</Keyword>
-    <ProjectGuid>{3a72b2f8-da95-4b32-b0ad-af31c87523f0}</ProjectGuid>
-    <RootNamespace>Source</RootNamespace>
+    <ProjectGuid>{3052527b-f682-465b-a93a-2259ce2b9193}</ProjectGuid>
+    <RootNamespace>TestProj</RootNamespace>
     <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
+    <ProjectName>Text Editor</ProjectName>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -127,6 +128,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="Auxiliary functions.h" />
+    <ClInclude Include="Buffer class.h" />
+    <ClInclude Include="Cursor class.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="Auxiliary Functions.cpp" />
+    <ClCompile Include="Buffer Class.cpp" />
+    <ClCompile Include="Cursor Class.cpp" />
     <ClCompile Include="Source.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
